@@ -5,7 +5,7 @@ description: Demonstrates how to use the flutter_uvc_camera plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: '>=2.19.0 <4.0.0'
+  sdk: '>=3.1.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -13,7 +13,7 @@ dependencies:
 
   flutter_uvc_camera:
     path: ../
-  path_provider: ^2.1.2
+
   cupertino_icons: ^1.0.2
 
 
