@@ -1,15 +1,9 @@
 library flutter_uvc_camera;
 
-import 'flutter_uvc_camera.dart';
-
-export 'src/camera_controller.dart';
-export 'src/camera_preview.dart';
-export 'src/models/camera_config.dart';
+export 'src/uvc_camera_controller.dart';
+export 'src/uvc_camera_view.dart';
 
 part 'src/models/uvc_camera_view_params_entity.dart';
 part 'src/models/uvc_camera_state.dart';
 part 'src/models/video_data_type.dart';
 part 'src/models/camera_preview_size_entity.dart';
-
-// Constants
-const String pluginVersion = '0.0.4';
