@@ -1,9 +1,3 @@
 part of flutter_uvc_camera;
 
-// 兼容旧版本的API
-@Deprecated('Use CameraState instead')
-enum UVCCameraState { 
-  opened, 
-  closed, 
-  error 
-}
+enum UVCCameraState { opened, closed, error }
